Rendering

+++ srcs/map.c
@@ -6,7 +6,7 @@
 /*   By: dienasci <dienasci@student.42sp.org.br >   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/08 20:39:15 by dienasci          #+#    #+#             */
-/*   Updated: 2021/10/21 13:48:37 by dienasci         ###   ########.fr       */
+/*   Updated: 2021/10/23 14:58:22 by dienasci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,7 +80,7 @@ t_list	*list_params(int fd)
 {
 	char	*line;
 	t_list	*list;
-	int		line_length;
+	size_t	line_length;
 
 	line = get_next_line(fd);
 	remove_breakline(line);
